Move to integration test
