+++ tests/integration_tests/test_text_splitter.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from langchain.text_splitter import CharacterTextSplitter
+from langchain.text_splitter import CharacterTextSplitter, TokenTextSplitter
 
 
 def test_huggingface_type_check() -> None:
@@ -21,3 +21,21 @@ def test_huggingface_tokenizer() -> None:
     )
     output = text_splitter.split_text("foo bar")
     assert output == ["foo", "bar"]
+
+
+class TestTokenTextSplitter:
+    """Test token text splitter."""
+
+    def test_basic(self) -> None:
+        """Test no overlap."""
+        splitter = TokenTextSplitter(chunk_size=5, chunk_overlap=0)
+        output = splitter.split_text("abcdef" * 5)  # 10 token string
+        expected_output = ["abcdefabcdefabc", "defabcdefabcdef"]
+        assert output == expected_output
+
+    def test_overlap(self) -> None:
+        """Test with overlap."""
+        splitter = TokenTextSplitter(chunk_size=5, chunk_overlap=1)
+        output = splitter.split_text("abcdef" * 5)  # 10 token string
+        expected_output = ["abcdefabcdefabc", "abcdefabcdefabc", "abcdef"]
+        assert output == expected_output
